refactor!: Rename "gpg" to "opengpg" (#13)
Complete the rename of the provider-name (in docs/examples/tests),
go-package-name, and resource-names from "gpg" to "opengpg".

Renaming the resource-name from "gpg_encrypted_message" to
"opengpg_encrypted_message" is a breaking change, so we want to do this
before anyone starts using the provider. Terraform's best practices
recommend using the provider's name as the prefix for resource-names:
https://developer.hashicorp.com/terraform/plugin/best-practices/naming#resource-names